GLEW now holds support for both 32 and 64bit versions. Qt only support 32bit, so we're forced to use that on this project...
Fixed rendering in OpenGL3GraphicsScene, which right now only uses OpenGL2. This is because the paint engine in Qt only support it, and is required to use QGraphicsView. Oh well, I'll be running i compatibility mode with 3.3 then :)

Last, added linkage/project inclusion to FactotumEDK, my property-driven, component-based entity system architecture. I'll be developing it as part of the project, and make it a central part of the client's architecture.